+++ src/components/CMakeLists.txt
@@ -8,22 +8,22 @@ base/i2c_controller.cpp
 base/i2c_target_communication_with_obc.cpp
 base/gpio_connection_with_obc.cpp
 
-aocs/gyro_sensor.cpp
-aocs/initialize_gyro_sensor.cpp
-aocs/magnetometer.cpp
-aocs/initialize_magnetometer.cpp
-aocs/magnetorquer.cpp
-aocs/initialize_magnetorquer.cpp
-aocs/reaction_wheel_ode.cpp
-aocs/reaction_wheel.cpp
-aocs/initialize_reaction_wheel.cpp
-aocs/reaction_wheel_jitter.cpp
-aocs/star_sensor.cpp
-aocs/initialize_star_sensor.cpp
-aocs/sun_sensor.cpp
-aocs/initialize_sun_sensor.cpp
-aocs/gnss_receiver.cpp
-aocs/initialize_gnss_receiver.cpp
+real/aocs/gyro_sensor.cpp
+real/aocs/initialize_gyro_sensor.cpp
+real/aocs/magnetometer.cpp
+real/aocs/initialize_magnetometer.cpp
+real/aocs/magnetorquer.cpp
+real/aocs/initialize_magnetorquer.cpp
+real/aocs/reaction_wheel_ode.cpp
+real/aocs/reaction_wheel.cpp
+real/aocs/initialize_reaction_wheel.cpp
+real/aocs/reaction_wheel_jitter.cpp
+real/aocs/star_sensor.cpp
+real/aocs/initialize_star_sensor.cpp
+real/aocs/sun_sensor.cpp
+real/aocs/initialize_sun_sensor.cpp
+real/aocs/gnss_receiver.cpp
+real/aocs/initialize_gnss_receiver.cpp
 
 real/cdh/on_board_computer.cpp
 real/cdh/on_board_computer_with_c2a.cpp
