+++ src/components/real/aocs/gyro_sensor.hpp
@@ -1,40 +1,40 @@
 /**
- * @file magnetometer.hpp
- * @brief Class to emulate magnetometer
+ * @file gyro_sensor.hpp
+ * @brief Class to emulate gyro sensor (angular velocity sensor)
  */
 
-#ifndef S2E_COMPONENTS_AOCS_MAGNETOMETER_HPP_
-#define S2E_COMPONENTS_AOCS_MAGNETOMETER_HPP_
+#ifndef S2E_COMPONENTS_REAL_AOCS_GYRO_SENSOR_HPP_
+#define S2E_COMPONENTS_REAL_AOCS_GYRO_SENSOR_HPP_
 
-#include <environment/local/local_environment.hpp>
+#include <dynamics/dynamics.hpp>
 #include <library/logger/loggable.hpp>
 #include <library/math/quaternion.hpp>
 
-#include "../base/component.hpp"
-#include "../base/sensor.hpp"
+#include "../../base/component.hpp"
+#include "../../base/sensor.hpp"
 
-const size_t kMagDim = 3;  //!< Dimension of magnetometer
+const size_t kGyroDim = 3;  //!< Dimension of gyro sensor
 
 /**
- * @class MagSensor
- * @brief Class to emulate magnetometer
+ * @class Gyro
+ * @brief Class to emulate gyro sensor
  */
-class MagSensor : public ComponentBase, public SensorBase<kMagDim>, public ILoggable {
+class Gyro : public ComponentBase, public SensorBase<kGyroDim>, public ILoggable {
  public:
   /**
-   * @fn MagSensor
+   * @fn Gyro
    * @brief Constructor without power port
    * @param [in] prescaler: Frequency scale factor for update
    * @param [in] clock_gen: Clock generator
    * @param [in] sensor_base: Sensor base information
    * @param [in] sensor_id: Sensor ID
    * @param [in] q_b2c: Quaternion from body frame to component frame
-   * @param [in] magnet: Geomagnetic environment
+   * @param [in] dynamics: Dynamics information
    */
-  MagSensor(const int prescaler, ClockGenerator* clock_gen, SensorBase& sensor_base, const int sensor_id, const libra::Quaternion& q_b2c,
-            const MagEnvironment* magnet);
+  Gyro(const int prescaler, ClockGenerator* clock_gen, SensorBase& sensor_base, const int sensor_id, const libra::Quaternion& q_b2c,
+       const Dynamics* dynamics);
   /**
-   * @fn MagSensor
+   * @fn Gyro
    * @brief Constructor with power port
    * @param [in] prescaler: Frequency scale factor for update
    * @param [in] clock_gen: Clock generator
@@ -42,15 +42,15 @@ class MagSensor : public ComponentBase, public SensorBase<kMagDim>, public ILogg
    * @param [in] sensor_base: Sensor base information
    * @param [in] sensor_id: Sensor ID
    * @param [in] q_b2c: Quaternion from body frame to component frame
-   * @param [in] magnet: Geomagnetic environment
+   * @param [in] dynamics: Dynamics information
    */
-  MagSensor(const int prescaler, ClockGenerator* clock_gen, PowerPort* power_port, SensorBase& sensor_base, const int sensor_id,
-            const libra::Quaternion& q_b2c, const MagEnvironment* magnet);
+  Gyro(const int prescaler, ClockGenerator* clock_gen, PowerPort* power_port, SensorBase& sensor_base, const int sensor_id,
+       const libra::Quaternion& q_b2c, const Dynamics* dynamics);
   /**
-   * @fn ~MagSensor
+   * @fn ~Gyro
    * @brief Destructor
    */
-  ~MagSensor();
+  ~Gyro();
 
   // Override functions for ComponentBase
   /**
@@ -72,17 +72,17 @@ class MagSensor : public ComponentBase, public SensorBase<kMagDim>, public ILogg
   virtual std::string GetLogValue() const;
 
   /**
-   * @fn GetMagC
-   * @brief Return observed magnetic field on the component frame
+   * @fn GetOmegaC
+   * @brief Return observed angular velocity of the component frame with respect to the inertial frame
    */
-  inline const libra::Vector<kMagDim>& GetMagC(void) const { return mag_c_; }
+  inline const libra::Vector<kGyroDim>& GetOmegaC(void) const { return omega_c_; }
 
  protected:
-  libra::Vector<kMagDim> mag_c_{0.0};            // observed magnetic field on the component frame [nT]
+  libra::Vector<kGyroDim> omega_c_{0.0};         //!< Observed angular velocity of the component frame with respect to the inertial frame [rad/s]
   int sensor_id_ = 0;                            //!< Sensor ID
   libra::Quaternion q_b2c_{0.0, 0.0, 0.0, 1.0};  //!< Quaternion from body frame to component frame
 
-  const MagEnvironment* magnet_;  //!< Geomagnetic environment
+  const Dynamics* dynamics_;  //!< Dynamics information
 };
 
-#endif  // S2E_COMPONENTS_AOCS_MAGNETOMETER_HPP_
+#endif  // S2E_COMPONENTS_REAL_AOCS_GYRO_SENSOR_HPP_
