+++ src/components/real/aocs/initialize_magnetometer.cpp
@@ -4,7 +4,7 @@
  */
 #include "initialize_magnetometer.hpp"
 
-#include "../base/initialize_sensor.hpp"
+#include "../../base/initialize_sensor.hpp"
 #include "library/initialize/initialize_file_access.hpp"
 
 MagSensor InitMagSensor(ClockGenerator* clock_gen, int sensor_id, const std::string fname, double compo_step_time, const MagEnvironment* magnet) {
