+++ src/components/real/aocs/initialize_magnetorquer.hpp
@@ -3,10 +3,10 @@
  * @brief Initialize functions for magnetorquer
  */
 
-#ifndef S2E_COMPONENTS_AOCS_INITIALIZE_MAGNETORQUER_HPP_
-#define S2E_COMPONENTS_AOCS_INITIALIZE_MAGNETORQUER_HPP_
+#ifndef S2E_COMPONENTS_REAL_AOCS_INITIALIZE_MAGNETORQUER_HPP_
+#define S2E_COMPONENTS_REAL_AOCS_INITIALIZE_MAGNETORQUER_HPP_
 
-#include <components/aocs/magnetorquer.hpp>
+#include <components/real/aocs/magnetorquer.hpp>
 
 /**
  * @fn InitMagTorquer
@@ -31,4 +31,4 @@ MagTorquer InitMagTorquer(ClockGenerator* clock_gen, int actuator_id, const std:
 MagTorquer InitMagTorquer(ClockGenerator* clock_gen, PowerPort* power_port, int actuator_id, const std::string fname, double compo_step_time,
                           const MagEnvironment* mag_env);
 
-#endif  // S2E_COMPONENTS_AOCS_INITIALIZE_MAGNETORQUER_HPP_
+#endif  // S2E_COMPONENTS_REAL_AOCS_INITIALIZE_MAGNETORQUER_HPP_
