+++ src/components/real/aocs/initialize_star_sensor.cpp
