+++ src/components/real/aocs/initialize_sun_sensor.cpp
