+++ src/components/real/aocs/initialize_sun_sensor.hpp
@@ -3,10 +3,10 @@
  * @brief Initialize functions for sun sensor
  */
 
-#ifndef S2E_COMPONENTS_AOCS_INITIALIZE_SUN_SENSOR_HPP_
-#define S2E_COMPONENTS_AOCS_INITIALIZE_SUN_SENSOR_HPP_
+#ifndef S2E_COMPONENTS_REAL_AOCS_INITIALIZE_SUN_SENSOR_HPP_
+#define S2E_COMPONENTS_REAL_AOCS_INITIALIZE_SUN_SENSOR_HPP_
 
-#include <components/aocs/sun_sensor.hpp>
+#include <components/real/aocs/sun_sensor.hpp>
 
 /**
  * @fn InitSunSensor
@@ -32,4 +32,4 @@ SunSensor InitSunSensor(ClockGenerator* clock_gen, int sensor_id, const std::str
 SunSensor InitSunSensor(ClockGenerator* clock_gen, PowerPort* power_port, int sensor_id, const std::string fname, const SRPEnvironment* srp,
                         const LocalCelestialInformation* local_celes_info);
 
-#endif  // S2E_COMPONENTS_AOCS_INITIALIZE_SUN_SENSOR_HPP_
+#endif  // S2E_COMPONENTS_REAL_AOCS_INITIALIZE_SUN_SENSOR_HPP_
