+++ src/components/real/aocs/magnetometer.cpp
