+++ src/components/real/aocs/star_sensor.cpp
